(#396) Move thumbnails management into 'Infra.Win32'
